chore: refactor Identities type
