+++ src/Application/StateIdentifierFactory.ts
@@ -1,4 +1,4 @@
-import { IdentifierBuilder, Identities } from "./types.ts";
+import { IdentifierBuilder, Identities } from "./Identities.ts";
 import { StateIdentifiers } from "../Domain/types.ts";
 
 export function stateIdentifierFactory(
@@ -10,7 +10,7 @@ export function stateIdentifierFactory(
   }
 
   const keys: StateIdentifiers = {
-    apiKey: identifierBuilder("user").ownedBy(identities.apiKey),
+    apiKey: identifierBuilder("apiKey").ownedBy(identities.apiKey),
   };
 
   if (identities.ip !== undefined) {
